refactor: move SMB specific test utils into smb_test.go
